requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
django-ninja==0.19.0
django-rest-knox==4.2.0
django-user-agents==0.4.0
djangorestframework==3.13.1
Pillow==9.2.0
pycparser==2.21
pydantic==1.9.1
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.3.0
ua-parser==0.15.0
user-agents==2.2.0
